Make SegmentProvider testable & add some tests
Make SegmentProvider testable & add some tests so that we can have some
coverage of what we are doing.

After looking at this it turns out that the way Flutter Segment packages
API is designed is such that it uses static methods and a singleton
pattern. This means that we can't really use Mockito to mock it out.
Beyond that in Dart you can't just pass a class around so we can't use
it as an implicit interface. Therefore, I created our own
`SegmentInterface` to represent the interface of the Segment library but
as an instance rather than static methods. This allowed me to then
implement an extremely thin `SegmentDriver` that implements the
`SegmentInterface`. Which in turn enables me to dependency inject a
`SegmentInterface` into the `SegmentProvider`.

This also facilitates us being able to define a SegmentDriverTestDouble
that we can then use to test the SegmentProvider in a state based
testing manner.

skip_ticket_id_check

<!-- ps-id: 6a09a894-5458-4762-89b6-493e5e95657e -->